+++ src/Compiler/Concerns/CompilesForwardedAttributes.php
@@ -16,7 +16,7 @@ trait CompilesForwardedAttributes
         $paramsToKeep = [];
 
         foreach ($node->parameters as $param) {
-            if ($this->isCacheParam($param)) {
+            if ($this->isCacheParam($param) || Str::startsWith($param->materializedName, $this->compilerDirectiveParams)) {
                 $compilerParams[] = $param;
 
                 continue;