+++ src/Runtime/Component.php
@@ -128,6 +128,16 @@ class Component extends AbstractComponent
         return $this->macros;
     }
 
+    public function hasSlot(string $slotName): bool
+    {
+        return $this->slots->hasSlot($slotName);
+    }
+
+    public function hasDefaultSlot(): bool
+    {
+        return $this->slots->hasDefaultSlotContent();
+    }
+
     public function __call($method, $parameters)
     {
         if (! isset($this->macros[$method])) {